# Performance improvement: simplified json.encode