--- a/lua/LUA/ak/core/CoreLuaModule.lua
+++ b/lua/LUA/ak/core/CoreLuaModule.lua
@@ -14,7 +14,7 @@ function CoreLuaModule.setRegisteredLuaModules(modules)
 end
 
 --- Diese Funktion wird einmalig durch ModuleRegistry.initTasks() aufgerufen
--- Ist ein Modul für EEPWeb vorhanden, dann solltes in dieser Funktion aufgerufen werden
+-- Ist ein Modul für EEPWeb vorhanden, dann sollte es in dieser Funktion aufgerufen werden
 -- @author Andreas Kreuz
 function CoreLuaModule.init()
     if not CoreLuaModule.enabled or initialized then