--- a/lua/LUA/ak/core/ModuleRegistry.lua
+++ b/lua/LUA/ak/core/ModuleRegistry.lua
@@ -127,9 +127,14 @@ function ModuleRegistry.runTasks(cycleCount)
     local t4 = os.clock()
 
     if ModuleRegistry.debug then
-        print(string.format("[ModuleRegistry] runTasks(cycleCount) time: %.0f ms " ..
-                                "(%.0f ms init, %.0f ms runTask, %.0f ms serveData)", (t4 - t1) * 1000,
-                            (t2 - t1) * 1000, (t3 - t2) * 1000, (t4 - t3) * 1000))
+        print(string.format("[ModuleRegistry] runTasks(%d) time: %.0f ms " ..
+                            "(%.0f ms initTasks, %.0f ms runTask, %.0f ms serveData)",
+                            cycleCount,
+                            (t4 - t1) * 1000,    -- total time
+                            (t2 - t1) * 1000,    -- initTasks
+                            (t3 - t2) * 1000,    -- runTask
+                            (t4 - t3) * 1000)    -- serveData
+        )
     end
 end