--- a/lua/LUA/ak/data/TrainsAndTracksJsonCollector.lua
+++ b/lua/LUA/ak/data/TrainsAndTracksJsonCollector.lua
@@ -1,5 +1,5 @@
 if AkDebugLoad then print("Loading ak.data.TrainsAndTracksJsonCollector ...") end
-
+local ServerController = require("ak.io.ServerController")
 
 TrainsAndTracksJsonCollector = {}
 local TrackCollector = require("ak.data.TrackCollector")
@@ -10,7 +10,27 @@ TrainsAndTracksJsonCollector.name = "ak.data.TrainsAndTracksJsonCollector"
 
 local data = {}
 
-local trackTypes = {"auxiliary", "control", "road", "rail", "tram"}
+-- Identitfy track types for which data should get collected
+local trackTypesAll = {"auxiliary", "control", "road", "rail", "tram"}
+local trackTypes = {}
+local activeEntries = ServerController.activeEntries
+if not activeEntries or next(activeEntries) == nil then -- empty list
+	-- All track types
+	trackTypes = trackTypesAll
+else	
+	-- Selected track types
+	for _, trackType in ipairs(trackTypesAll) do
+		for key, value in pairs(activeEntries) do
+			if value then
+				if string.find(key, trackType) then
+					table.insert(trackTypes, trackType)
+					break
+				end
+			end
+		end
+	end
+end
+
 local trackCollectors = {}
 do
     for _, trackType in ipairs(trackTypes) do