I still try to export data from running EEP within less time than 200 ms even for large railroad plans to get at least a frame rate of 5/sec for my external program.

Concatenating large amounts of strings is quite slow in Lua.
It's much faster to collect strings as they are and use `table.concat` to produce the final result within one step.
Constructing strings using `string.format` is faster than concatenating several parts using multiple `..` operators.

The standard json-library used so far is safe (it performs cycle checks) and robust (it accepts any data for key and data fields) but slow.

I've replaced it which a simplified but faster version. 

Test for railroad plan "Enkel_ANLAGE_Flughafen" (if I activate all data collector):

`json.lua`
time is 5123 ms --- waitForServer: 0 ms, initialize: 1600 ms, commands:  0 ms, collect: 236 ms,  expand:  13 ms  encode: 3193 ms  write:  81 ms
time is 3544 ms --- waitForServer: 0 ms, initialize:    0 ms, commands:  0 ms, collect: 215 ms,  expand:  11 ms  encode: 3217 ms  write: 101 ms
time is 3961 ms --- waitForServer: 0 ms, initialize:    0 ms, commands:  0 ms, collect: 308 ms,  expand:  18 ms  encode: 3540 ms  write:  95 ms
time is 4456 ms --- waitForServer: 0 ms, initialize:    0 ms, commands:  0 ms, collect: 282 ms,  expand:   9 ms  encode: 4064 ms  write: 101 ms
time is 4821 ms --- waitForServer: 0 ms, initialize:    0 ms, commands:  1 ms, collect: 343 ms,  expand:  17 ms  encode: 4365 ms  write: 95 ms

`jsonEncode.lua`
time is 5007 ms --- waitForServer: 0 ms, initialize: 2444 ms, commands:  0 ms, collect: 312 ms,  expand:   9 ms  encode: 2127 ms  write: 115 ms
time is 2512 ms --- waitForServer: 0 ms, initialize:    0 ms, commands:  1 ms, collect: 287 ms,  expand:  18 ms  encode: 2108 ms  write:  98 ms
time is 2844 ms --- waitForServer: 0 ms, initialize:    0 ms, commands:  0 ms, collect: 280 ms,  expand:  14 ms  encode: 2451 ms  write:  99 ms
time is 3193 ms --- waitForServer: 0 ms, initialize:    0 ms, commands:  0 ms, collect: 298 ms,  expand:  12 ms  encode: 2775 ms  write: 108 ms
time is 3363 ms --- waitForServer: 0 ms, initialize:    0 ms, commands:  0 ms, collect: 348 ms,  expand:  28 ms  encode: 2847 ms  write: 140 ms
time is 3205 ms --- waitForServer: 0 ms, initialize:    0 ms, commands:  0 ms, collect: 333 ms,  expand:  29 ms  encode: 2723 ms  write: 120 ms

-> We save ca 1/3 of the time required for the encode step.


-- 

_Well this is just a first step. Actually I do not use any json-encoding any more in my own project but produce the string table containing complete records instead of multiple entries for data fields directly during data collection._